move get_task_status to base
